Changes on the sidebar - components separate
