+++ admin-site/src/components/Sidebar/Types.ts
@@ -1,0 +1,6 @@
+export interface MenuItem {
+  label: string;
+  icon?: React.ReactNode;
+  path?: string;
+  children?: MenuItem[];
+}
